:wrench:

+++ package/cpp/rnwgpu/api/GPUComputePassEncoder.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <memory>
+#include <optional>
 #include <variant>
 #include <vector>
 
@@ -7,7 +9,6 @@
 
 #include "RNFHybridObject.h"
 
-#include "ArrayBuffer.h"
 #include "AsyncRunner.h"
 
 #include "webgpu/webgpu_cpp.h"
@@ -37,7 +38,7 @@ public:
   void setBindGroup(
       uint32_t index,
       std::variant<std::nullptr_t, std::shared_ptr<GPUBindGroup>> bindGroup,
-      std::optional<std::vector<double>> dynamicOffsets);
+      std::optional<std::vector<uint32_t>> dynamicOffsets);
 
   std::string getLabel() { return _label; }
